+++ f2/package.json
@@ -68,7 +68,8 @@
     "check-translations": "node utils/checkTranslations.js",
     "upgrade-interactive": "npm-check --update",
     "storybook": "start-storybook -p 9009 -s public",
-    "build-storybook": "build-storybook -s public"
+    "build-storybook": "build-storybook -s public",
+    "extract-fields": "node utils/extractFields.js"
   },
   "cypress-cucumber-preprocessor": {
     "nonGlobalStepDefinitions": true
